backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
greenlet==3.2.4

# LLM Integration
anthropic==0.7.8
pydantic==2.5.0

# Additional utilities
python-dotenv==1.0.0
jinja2==3.1.2

# Web interface and templating
aiofiles==23.2.1
python-dateutil==2.8.2

# Enhanced validation and utilities
email-validator==2.1.0
typing-extensions==4.8.0

# Authentication
PyJWT==2.8.0
cryptography==41.0.7
passlib[bcrypt]==1.7.4

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
